requirements.txt
Django
tink[gcpkms]
cryptography
google-cloud-kms
